requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.66.0
